feat: implement modular UI components for persona management including list rendering, global settings, and native element relocation
